The Three Body Problem.
I really liked this book, but I felt like it fell off a bit towards the
end. I thought the setting was unique, and that they did an excellent
job with the game and the initial contact. The idea of having 400 years
until arrival was neat as well. However, the whole two proton thing was
a bit much given the relatively "harder" sci-fi aspects throughout the
book. Not quite sure how the author would be able to write himself out
of the amazing hook in the opening chapters wihtout it though...