The Billion Dollar Molecule
Great read - one of my top this year. Pretty accurate insights into academic
research (even though it's mainly about a startup) and the politics involved.
Amazing how timeless many of the startup tropes in this book are considering it
was written in the early 90's. I'm interested in reading the follow-up now that
Vertex is a bonafide big-pharma company.