Odalisque
Finally finished Vol. 1 of the Baroque Cycle. Overall, I think King of
the Vagabonds was the best. Odalisque, like Quicksilver, was mostly
exposition and world building. There is so much depth and detail in all
of these books that a second (or third) reading is probably needed to
appreciate it all. Knowing Stephenson, he's probably encrypted some hidden
messages in the letters between Leibniz and Eliza. I do intend to finish
the entire series, although given the length, I'm not sure when this
will happen...