The Gene
Solid history of genetics/genomics. I burned through this since much of it was
review. Would highly reccommend for non-scientists to get a sense of the
genetics/genomics field. I think the author does a good job at touching
on the heavy social and ethical implications inherent to this work.
I did feel that he skimped on the next-generation sequencing revolution,
which actually enabled much of the advances over the last 10 years.