Zero to One
This book is conflicting. On one hand, Zero to One manages to compress
volumes of startup advice into a straight-forward and concise
manuscript. If you've lurked YC for any period of time, its abundantly
clear how much of Thiel's ideas have penetrated the cultures ethos.
However, some of Thiel's remarks make it hard to take the rest of the
work seriously (e.g. "never fund a tech CEO in a suit" or all biotech
workers are over-paid, mindless drones).

Having said that, I think this is a great book to get someone quickly up
to speed with the current way to start a startup. For those who want to
dive deeper, there's many threads to go on:
 - Stagnation Theory (see HyperNormalization/concrete view of future)
 - Fat Tails/Black Swans (shout out to Taleb)
 - Monopolies are good (would be interesting to read a more in-depth
   economic analysis of this statement)