Meditations
Read this book during a train and plane ride, although I would not
reccomend reading it in large chunks. Found it to be chock-ful
of wisdom. It's amazing that this work has survived thousands of years
and still manages to be relevant today. Some of Marcus' issues (getting
out of bed in the morning) are truly timeless.

I think one can make a clear link between Marcus' writings and Terror
Management Theory (I may need to follow this up with Denial of Death).
He constantly grapples with the inevitability of death and fleeting
nature of life in order to motivate his actions. There's also surprising
overlaps with Mindfulness Meditation.