The Name of the Rose
One of the more challenging books I've read to date. The way I like to
think of Eco's writing is a mix between the prose of Nabokov and geekdom
of Stephenson. Instead of getting a multi-page tangent on cryptography or
orbital mechanics, Eco gives the reader a lesson in ancient Christian
sects or lengthy debates on whether Jesus laughed. While the writing is
beautiful, I personally find the subjects to be a bit dry...

Overall, I appreciated the work, but found the pacing and tangents to be
a bit tiring. I also feel that I missed an entire portion of the work by
not taking the time to translate the Latin/Spanish/German/Greek phrases.
I certainly feel that this novel requires multiple readings to fully
digest. Whether or not I'm willing to do so is another question
entirely...