Science Business
Much ink has been spilled on the subject of the pharmaceutical
industry's productivity issue. Science Business was one of the first to
offer any sort of rationale as to why drug companies struggle to get
treatments to market in a timely and cost-effective manner. Given that
there hasn't been any major advances in productivity over the past 10
years since this book was published, it's safe to say that this is still
a major issue.

I think the author does a good job at outlining what goes in to
developing a drug, and why it is difficult to do so. Although some of
the technologies are a little dated, the fundemental idea of unknown
unknowns still plauges biological research. I also enjoyed Pisano's
outline of the three major eras of biotech, leading up ~2007. This trend
is still apparent with the rise of immunotherapy and the rebirth of gene
therapy recently.

In terms of solutions, I'm not sure Pisano hit the mark. His advice
seemed overly generic - better collaborations between startups and big
pharma, better integration of interdisciplenary research, and less
restrictive university licensing (amongst others). Many of his solutions
are in place in some form today, and there is still an issue.
Personally, I think we are still somewhat immature in our understanding
of the fundamental biology that underlies the drug discovery process
(although we are certainly getting better!).

Overall, I would reccomend this as a quick read (~200 pages) to get up
to speed with  some of the common thinking around how drug development
is performed, and the problems in the industry as a whole. I wouldn't
take Pisano's word as gospel, though.