If on a winter's night a traveler
I feel like I've been doing meta reading these past couple of months...
Every piece of fiction I pick up is a book about books. In this case,
If on a Winter's Night a Traveller is made up of multiple books, each
linked together by a story narrated in the second person.

The book opens with "You are about to begin reading Italo Calvino’s new
novel..." I initially thought this was an introduction from another
author. It wasn't until the third chapter that I understood the
narrative structure. Needless to say, I was hooked.

Each of the stories is masterfully crafted to draw you in, even when you
know you will never know how they are resolved. Towards the end of the
novel, I found myself getting frustrated. Not so much at the repetitive
nature of the plot structure, more at the fact that I wouldn't be able to
finish these stories (although I must admit that many of them seem like
the sort of guilty pleasure $1 romance novels).

*SPOILERS*
My only major gripe is that the story falls flat at the end. The
international conspiracy was a bit far-fetched, and the abrupt marriage
of The Reader and Ludmilla seemed unmotivated. Also, I'm not sure this
book would hold up to multiple readings, given the charm of the novel
is the stories contained within it.