Nothing Is True and Everything Is Possible
I initially got this book as I was interested in learning more about
Vladislav Surkov. Although he was only mentioned briefly, the rest of the
book reveals the overall structure of the Russian PR machine through
a series of extended profiles and an interwoven narrative. I found some
of the points particularly poignant, especially now that we are in the
"fake news era."