Mindfulness in Plain English
This book is one of the most clear, concise, and easy to read pieces of nonfiction
I have ever come across. Bhante Gunaratana manages to convey the nuts
and bolts of Mindfulness in a light and conversational manner that is
surprisingly secular. In addition to being practical guide, there are deeper
and more insightful sections that are worth returning to. I would highly
recommend this book to anyone who is interested in learning more about meditation.