The Confusion
Neil Stephenson is a demanding author. Take a random review of one of
his novels and you are almost guaranteed to find words like: "long,"
"boring," and "infodump." I think these descriptions are mostly correct
- the first volume of the Baroque Cycle was practically 1000 pages of
world building.

The Confusion, volume two of the Baroque Cycle, is a perfect example of
why Stephenson fans find his work so enjoyable. The characters are let
loose in series of adventures that span the entire globe over 20 years.
As the pieces of the story started to fall into place, I felt a strong
sense of delayed gratification (or potentially, a powerful bout of
cognitive dissonance).

Of the two stories con-fused together, I found Jack's to be much more
engaging. Daniel and Eliza's story seemed muted at times, but the
political/economic drama that they participate in lends itself to a
softer touch.

Like Quicksilver before it, I am a little hesitant to dedicate the time
and effort to finish the next volume. Hopefully, The System of the World
continues the upward trend set by The Confusion, but knowing
Stephenson's penchant for botching endings, I'm approaching it with
cautious optimism.