Hackers and Painters
If I would have known that Hackers and Painters was just a printed out
version of some of Paul Graham's early essays, I probably wouldn't have
bought it. That being said, I'm glad I read (and re-read) these essays.
PG has a clear and conversational writing style that I admire, and is a
wealth of interesting insights. I thought the chapters on taboos and
hacker culture were particularly compelling, but the musing about
programming languages and spam filtering felt out of place. The
unorganized nature of the essays, combined with the fact that they are
readily available on his website left me slightly disappointed.