Man's Search for Meaning
If you look at any list of "books that will change your life" Man's
Search for Meaning is bound to be on it. Having read it, I can say that
it has not (unfortunately) changed my life, but is certainly a profound
read.

The first half of the novel is dedicated to Frankl's experience as a
prisoner in various concentration camps during WWII. His accounts are
gruesome and difficult to read. I had a hard time getting through them
as it reminded me of similar stories I heard about my Grandfather's
experience.

The second half centers around logotherapy, the brand of psychology
Frankl developed before, during, and after the war. I think there are
some useful tid-bits here and there. Two lines in particular really
stood out to me. The first is the Nietzsche quote: "He who has a why
can bear almost any how." The second is: "Everything can be taken from a
man but one thing; the last of the human freedoms — to choose one's
attitude in any given set of circumstances, to choose one's own way."