Post Office
Reading Post Office felt like watching a train wreck. It was arguably
one of the most depressing novels I've ever read, but I found myself
unable to put it down. The way Bukowski manages to capture the
loneliness, isolation, and self-hatred of a drunk is made even more
poignant by the fact that Post Office is practically an autobiography.
Beyond the voyeuristic appeal of stepping into the life of a scumbag
lies a deep commentary on the crushing absurdities and endless
bureaucracy of most jobs that only adds to the feelings of despair.