Spew
Spew is the the tip of the Stephenson iceburg. It's short, punchy, and distils
his work into an unusually (for Stephenson) short work. He paints an erily
familiar future where massive ad-tech conglomerates have developed
hyper-specific ("24 year-old single white female New Derisive with
post-Disillusionist leanings") profiles to better target content to people, and
are constantly on the look out for "unexploited market niches of cavernous
proportions (1/100th of 1% of the population)." I would reccomend this,
followed by Snow Crash for those looking to get into one of my favorite authors.

Invisible Cities

Quintisential Calvino. It's amazing what he can do with such a paucity of
words. I would reccommend against reading this book in a single sitting. It's
better to read a few cities and let them digest. If possible, get a paper copy
as there is a higher order structure to the ordering of the stories that you
can't take advantage of on the Kindle.

Metamagical Themas: Questing for the Essence of Mind and Pattern

In classic Hofstadter fashion, Metamagical Themas is a tome packed full of
ideas, self reference, puns, math, and music. I found it to be much more
accessible than Godel Escher Bach, while being equally enlightening. It amazes
me how current Hofstadter's commentary on AI, conciousness, and even gender
pronouns feel, considering the book was written over 30 years ago. If you are
at least tangentially interested in any of those subjects I highly recommend
reading this work. If 850 pages of 6x9 card stock is too much for you, I would
pick and choose from his Scientific American columns.

Never Let Me Go

I was inspired to read this after Ishiguro won the Nobel. Like Post Office,
Never Let Me Go is a trainwreck of a novel. You follow the lives of three young
students as they grow up in an elite boarding school in England called
Hailsham. As the reader, it is easy to divine their fate, but the students
remain frustratingly ignorant and needlessly positive even when faced with the
truth. I remember not liking this novel when I first finished it, but in time,
I've grown to appreciate it. I think the connection between the Hailsham
students' situation and our attitude about our own mortality is profound. This
connection is explained much more elegantly than me by Theo Tait - "Gradually,
it dawns on the reader that Never Let Me Go is a parable about mortality. The
horribly indoctrinated voices of the Hailsham students who tell each other
pathetic little stories to ward off the grisly truth about the future – they
belong to us; we've been told that we're all going to die, but we've not really
understood."

The Southern Reach Trilogy

I will forgo the plot summary in this review to avoid spoilers. I think these
books are worth diving into without knowing much. One of the most common
criticism of the series is that it Annihilation is a knock off of Roadside
Picnic. However having read both of them, I think that they both deal with
similar themes but are different (and good!) enough to warrant a read.

The Beginning of Infinity

The Beginning of Infinity by David Deutsch is a book about the progress of
humanity. The central argument is that the Scientific Revolution gave us a tool
to unlock infinite progress - namely a systematic method for developing good
explanations. Good explanations are descriptions of our universe that are hard
to vary, are often elegant, and have reach. For example, Newton's Laws were
good explanations because they were robust to centuries of criticism, were
certainly elegant, affected many aspects of our lives. Furthermore, the
predictions about our universe from Newton's Laws are largely recapitulated by
later (better) explanations such as Einstein's Theory of Relativity. With this
framework in place, Deutsch goes on to explain a variety of subjects including
objective beauty, the evolution of memes, artificial intelligence, and the
multiverse.

Overall, I really enjoyed this book, but found it a little overwrought. Deutsch
often formalizes his arguments in such a way that it feels like a derivation
from first principles. I found this a little excessive and tiring after I
understood the larger points. I also wish that there was a synonym for
parochial, which seems to work its way into every paragraph. Having said that
it's an excellent book with many interesting connections to previous works I've
read (Hackers and Painters/The Baroque Cycle).