Catching up...
The Turkish Gambit

Cute novel I found in some Hacker News thread. I enjoyed the setting but other
than that, it wasn't very memorable.

Ancillary Justice

A compelling start to an expansive space opera. The lack of gendered pronouns
was a fun device that appropriate for the main character (an omnipresent AI who
has gone rogue). Cool ideas about consciousness, AGI, and gender.

Why Buddhism is True

One of these books that would be better as an essay. Fortunately, this is
conveniently provided in Appendix A. It doesn't provide much outside of the
standard secular meditation talking points, but might be interesting to someone
with no exposure to the concepts.

The Book of Why

A fascinating look into the history of Causal Inference from a giant in the
field. It's interesting and sad to think what might have been if causal
inference gained traction earlier on. I thoroughly enjoyed Pearl's skeptical
rumination on AI and his unique take on the Montey Hall problem. One of my
favorite books of 2018.

A Confederacy of Dunces

Hilarious and scathing account of Ignatius J. Reilly, an emblematic
fail-son/neckbeard. Another top read of 2018.

Dark Forest

Grander, more thought-provoking, and entertaining than the first. The Dark
Forest theory - communication is dangerous and will likely lead to destruction,
is harrowing and plausible. As is the depiction of first contact. If you are
uncertain about finishing the trilogy after reading the first, press on.

Bad Blood

When move fast and break things goes too far.

Flatland

Catches a lot of flack for the blatant sexism (possibly addressed in the
preface?), but looking past that provides a deeper insight into how we perceive
the world. The analogy of a 3D object projected into a 2D world provided some
insight into how a higher dimensional object projected into our world would
look.

The Prince

Didn't live up to its infamous reputation. Possibly because many of the ideas
have pervaded popular culture. Seems like the precursor to the personal
advice/how to run a business genre.

Los Angeles: The Architecture of Four Ecologies

Should be required reading for any long-time Angeleno. A love letter to LA that
explains how LA came to be.

Industrial Society and Its Future

Came to this through Bill Joy's excellent "Why the Future Doesn't Need Us."
Surprisingly poignant and prescient given the recent (albeit tepid) backlash
against big tech and rise of automation. Even his characterization of Leftists
feels as if it could be pulled from Breitbart. While, I think there is some
merit to his argument that technology has made life unfulfilling and is causing
widespread psychological suffering (just look at the critiques at social
media). I doubt returning to a hunter-gatherer society would be better...

The Duel

Insight into the life of Russian aristocrats during the late 1800's with a
commentary on honor, masculinity, envy, and the dangers of ideology.

Siddhartha

A book about life that should be re-read over the course of one's life.
Beautiful.

Of Human Bondage

In a similar vein to Siddhartha, Of Human Bondage follows the life of Philip
Carey. A brutal exploration of self discovery, relationships, obsession, and
existentialism. I wish I would have read this at a younger age, but wonder what
I would have taken from it.

Fortune's Formula

A fun history of the Kelly Criterion - a fabulously simple formula for sizing
bets with profound ramifications for the stock market. Also offers an insight
into how the best hedge funds exploit fundamental pricing discrepancies.

The Master and Margarita

A fantastic (in the classical sense) story set in Stalinist Russia that
operates on multiple levels. I feel like I just scratched the surface (which is
entertaining enough), but would be rewarded if I had a better understanding of
life in Russia at the time. Deserving of its praise.

Less Than Zero

Brett Easton Ellis certainly deserves his reputation for writing deeply
disturbing works. A nihilistic, indulgent look into the life of spoiled LA
youth.

The Great Stagnation

Huge fan of Tyler's work. He brings up the point that besides information
technology (and more recently biology), there has been limited progress in
basically every area of our life. It's definitely worth thinking about how the
things we interact with on a daily basis are changing and if they are
fundamentally different than 10, 20, 30 years ago. For a more concentrated
version of this, see the Founder's Fund manifesto.

Hunger

Considered to be one of the first "modern" novels. An interesting and
semi-autobiographical account of a truly starving artist that slides into
insanity.

Death's End

A deserving conclusion to the Three Body Problem trilogy. Loved how Liu handled
time and astronomical scale. The entire trilogy is some of the best Sci-Fi I've
read to date.

Super Thinking

Expansive survey of the most common mental models. This should be required
reading for young people. Amazing amount of wisdom packed into an easy to
digest format. It's a mile wide and an inch deep but serves a great jumping off
point for how to think about thinking.

Foucault's Pendulum

One of the more challenging books I've ever read - endless Stephenson-esque
diversions into medieval occult esoterica, numerous different languages, and
SAT words galore. It's like being stuck in the Museum of Jurassic Technology.
Coming out the other end, however you'll gain a deeper appreciation for the
dangers of irony and the power of skepticism. Particularly apropos given the
rise of Fake News. That said, I enjoyed The Name of the Rose much more.